bower, leaflet markercluster
used the leaflet markercluster plugin to draw results of test data
passed as a feature collection. we also implemented bower for package
managment, but copied the js code into our libs directory for now.
eventually we'll want to mange all dependencies for the frontend with
bower to make things easier to keep cleaned and maintained.